+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/validation/listing/reviewer/edition2014/AmbulatoryRequiredTestToolReviewer.java
@@ -33,8 +33,14 @@ public class AmbulatoryRequiredTestToolReviewer implements Reviewer {
                         && certRules.hasCertOption(cert.getNumber(), CertificationResultRules.TEST_TOOLS_USED)
                         && !ValidationUtils.containsCert(cert, TEST_TOOL_CHECK_CERTS)
                         && (cert.getTestToolsUsed() == null || cert.getTestToolsUsed().size() == 0)) {
-                    listing.getErrorMessages()
-                    .add(msgUtil.getMessage("listing.criteria.missingTestTool", cert.getNumber()));
+                    if (listing.getIcs() != null && listing.getIcs().getInherits() != null
+                            && listing.getIcs().getInherits().booleanValue()) {
+                        listing.getWarningMessages()
+                        .add(msgUtil.getMessage("listing.criteria.missingTestTool", cert.getNumber()));
+                    } else {
+                        listing.getErrorMessages()
+                        .add(msgUtil.getMessage("listing.criteria.missingTestTool", cert.getNumber()));
+                    }
                 }
             }
         }